backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.16
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
sqlalchemy==2.0.23
python-multipart==0.0.6
slowapi==0.1.9